Attempt to change the color in About-Form
Changed in line 10 Color from clBtnFace to clBtnShadow. This should change the standard grey to a slightly darker grey. As the shadow underneath a element to generate 3D feeling. The value clBtnFace is a defined system color and windows generates the shadowed version. As I read, clBtnFace is the color set in the system for 3D system element surfaces.